+++ test/dao.test.js
@@ -57,7 +57,9 @@ describe(Support.getTestDialectTeaser("DAO"), function () {
           expect(u2.username).to.equal(bio)
           done()
         })
-      })
+      }).error(function(err) {
+	done(err);
+      });
     })
   })
 
@@ -853,7 +855,7 @@ describe(Support.getTestDialectTeaser("DAO"), function () {
         expect(err).to.be.instanceof(Object)
         expect(err.validateTest).to.be.instanceof(Array)
         expect(err.validateTest[0]).to.exist
-        expect(err.validateTest[0]).to.equal('Validation isInt failed')
+	expect(err.validateTest[0].message).to.equal('Validation isInt failed')
         done()
       })
     })
@@ -866,7 +868,7 @@ describe(Support.getTestDialectTeaser("DAO"), function () {
         expect(err.validateCustom).to.exist
         expect(err.validateCustom).to.be.instanceof(Array)
         expect(err.validateCustom[0]).to.exist
-        expect(err.validateCustom[0]).to.equal('Length failed.')
+	expect(err.validateCustom[0].message).to.equal('Length failed.')
         done()
       })
     })
@@ -879,7 +881,7 @@ describe(Support.getTestDialectTeaser("DAO"), function () {
           expect(err.validateTest).to.exist
           expect(err.validateTest).to.be.instanceof(Array)
           expect(err.validateTest[0]).to.exist
-          expect(err.validateTest[0]).to.equal('Validation isInt failed')
+	  expect(err.validateTest[0].message).to.equal('Validation isInt failed')
           done()
         })
       })
@@ -1048,6 +1050,44 @@ describe(Support.getTestDialectTeaser("DAO"), function () {
           })
         })
       })
+    })
+  })
+  describe('many to many relations', function() {
+    var udo;
+    beforeEach(function(done) {
+      var self = this
+      this.User = this.sequelize.define('UserWithUsernameAndAgeAndIsAdmin', {
+        username: DataTypes.STRING,
+        age:      DataTypes.INTEGER,
+        isAdmin:  DataTypes.BOOLEAN
+      }, {timestamps: false})
+
+      this.Project = this.sequelize.define('NiceProject',
+        { title: DataTypes.STRING }, {timestamps: false})
+
+      this.Project.hasMany(this.User)
+      this.User.hasMany(this.Project)
+
+      this.User.sync({ force: true }).success(function() {
+        self.Project.sync({ force: true }).success(function() {
+          self.User.create({ username: 'fnord', age: 1, isAdmin: true })
+            .success(function(user) {
+              udo = user
+              done()
+            })
+        })
+      })
+    })
+    it.skip('Should assign a property to the instance', function(done) {
+      // @thanpolas rethink this test, it doesn't make sense, a relation has
+      // to be created first in the beforeEach().
+      var self = this;
+      this.User.find({id: udo.id})
+        .success(function(user) {
+          user.NiceProjectId = 1;
+          expect(user.NiceProjectId).to.equal(1);
+          done();
+        })
     })
   })