+++ test/hooks.test.js
@@ -464,7 +464,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('should return an error based on user', function(done) {
                 this.User.create({mood: 'happy'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal([ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ])
+		  expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                   done()
                 })
               })
@@ -501,7 +501,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('should return an error based on the hook', function(done) {
                 this.User.create({mood: 'happy'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal([ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ])
+		  expect(err.mood[0].message).to.equal( 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' )
                   done()
                 })
               })
@@ -552,7 +552,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('should return an error based on user', function(done) {
                 this.User.create({mood: 'happy'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal([ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ])
+		  expect(err.mood[0].message).to.equal( 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' )
                   done()
                 })
               })
@@ -601,7 +601,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('should return an error based on the hook', function(done) {
                 this.User.create({mood: 'happy'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal([ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ])
+		  expect(err.mood[0].message).to.equal( 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' )
                   done()
                 })
               })
@@ -698,7 +698,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
                 it('should return the user from the callback', function(done) {
                   this.User.create({mood: 'happy'}).error(function(err) {
                     expect(err).to.be.instanceOf(Error);
-                    expect(err.mood).to.deep.equal([ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ])
+		    expect(err.mood[0].message).to.equal( 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' )
                     done()
                   })
                 })
@@ -720,7 +720,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
                 it('should return the error without the user within callback', function(done) {
                   this.User.create({mood: 'happy'}).error(function(err) {
                     expect(err).to.be.instanceOf(Error);
-                    expect(err.mood).to.deep.equal([ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ])
+		    expect(err.mood[0].message).to.equal( 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' )
                     done()
                   })
                 })
@@ -835,7 +835,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('#create', function(done) {
                 this.User.create({mood: 'creative'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal( [ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ] )
+		  expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                   done()
                 })
               })
@@ -859,7 +859,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('#create', function(done) {
                 this.User.create({mood: 'happy'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal( [ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ] )
+		  expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                   done()
                 })
               })
@@ -940,7 +940,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
 
               this.User.create({mood: 'happy'}).error(function(err) {
                 expect(err).to.be.instanceOf(Error);
-                expect(err.mood).to.deep.equal( [ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ] )
+		expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                 done()
               })
             })
@@ -954,7 +954,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
 
               this.User.create({mood: 'happy'}).error(function(err) {
                 expect(err).to.be.instanceOf(Error);
-                expect(err.mood).to.deep.equal( [ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ] )
+		expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                 done()
               })
             })
@@ -1061,7 +1061,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('#create', function(done) {
                 this.User.create({mood: 'creative'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal( [ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ] )
+		  expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                   done()
                 })
               })
@@ -1085,7 +1085,7 @@ describe(Support.getTestDialectTeaser("Hooks"), function () {
               it('#create', function(done) {
                 this.User.create({mood: 'happy'}).error(function(err) {
                   expect(err).to.be.instanceOf(Error);
-                  expect(err.mood).to.deep.equal( [ 'Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral' ] )
+		  expect(err.mood[0].message).to.equal('Value "ecstatic" for ENUM mood is out of allowed scope. Allowed values: happy, sad, neutral')
                   done()
                 })
               })